+++ app/Sao/A2.php
@@ -6,6 +6,7 @@ use Facebook\WebDriver\Firefox\FirefoxOptions;
 use Facebook\WebDriver\Firefox\FirefoxProfile;
 use Facebook\WebDriver\Remote\DesiredCapabilities;
 use Facebook\WebDriver\Remote\RemoteWebDriver;
+use Facebook\WebDriver\WebDriverBy;
 use Illuminate\Support\Facades\Log;
 use Intranet\Componentes\Mensaje;
 use Intranet\Exceptions\CertException;
@@ -100,7 +101,7 @@ class A2
     public static function downloadFilesFromFcts(RemoteWebDriver $driver, $fcts, $certFile=null)
     {
         $signat = false;
-        $a1 = $a2 = $a3 = $fA1 = false;
+        $a1 = $a2 = $a3 = $fA1 = $a5 = false;
 
         foreach ($fcts as $fct) {
             if ($fct === 'FA1'){ //A1 forçat
@@ -114,6 +115,9 @@ class A2
             }
             if ($fct === 'A3') { //A3 forçat
                 $a3 = true;
+            }
+            if ($fct === 'A5') {
+                $a5 = true;
             }
             $fctAl = AlumnoFct::find($fct);
             if ($fctAl){
@@ -129,6 +133,9 @@ class A2
                 // Anexe 3
                 if ($a3 && $certFile){
                     self::annexe23($fctAl, $driver, $certFile,3);
+                }
+                if ($a5) {
+                    self::annexe5($fctAl, $driver, $certFile);
                 }
             }
         }
@@ -163,7 +170,9 @@ class A2
                 copy($tmpFile, $saveFile);
             }
             Firma::saveIfNotExists($annexe, $fctAl->idSao);
-            unlink($tmpFile);
+            if (file_exists($tmpFile)){
+                unlink($tmpFile);
+            }
             $driver->get(self::HTTPS_FOREMP_EDU_GVA_ES_INDEX_PHP_OP_2_SUBOP_0);
             sleep(1);
         }
@@ -220,5 +229,70 @@ class A2
             sleep(1);
         }
         return false;
+    }
+
+    private static function annexe5($fctAl, RemoteWebDriver $driver,$certFile):bool
+    {
+        $tmpDirectory = config('variables.shareDirectory')??storage_path('tmp/');
+        $doc = $fctAl->Fct->dual ? '201' : '101';
+        $annexe = $fctAl->Fct->dual ? 'A5DUAL' : 'A5';
+        $idSao = $fctAl->idSao;
+        $tmpFile = $tmpDirectory.$annexe.".pdf";
+        $saveFile = $fctAl->routeFile($annexe);
+        $x = config("signatures.files.".$annexe.".owner.x");
+        $y = config("signatures.files.".$annexe.".owner.y");
+        try {
+            $driver->get("https://foremp.edu.gva.es/index.php?accion=7&idFct=$idSao");
+            sleep(1);
+            $enlace = $driver->findElement(WebDriverBy::xpath("//a[contains(@class, 'enlDocFCT') and contains(., 'Informe de Consecución de Competencias (Autorrellenable)')]"));
+            $enlace->click();
+            sleep(1);
+            $printButton = $driver->findElement(WebDriverBy::xpath("//button[contains(.,'Imprimir documento')]"));
+            $printButton->click();
+            sleep(1);
+            self::waitForFile($tmpFile, 5);
+
+            // Copia el fitxer descarregat al destí
+            if (file_exists($tmpFile)) {
+                copy($tmpFile, $saveFile);
+            }
+
+        } catch (\Throwable $exception) {
+            Alert::danger($exception->getMessage());
+            if (file_exists($tmpFile)) {
+                copy($tmpFile, $saveFile);
+            }
+        } finally {
+
+            if ($certFile) {
+                DigitalSignatureService::sign(
+                    $tmpFile,
+                    $saveFile,
+                    $x,
+                    $y,
+                    $certFile
+                );
+                Firma::saveIfNotExists($annexe, $fctAl->idSao, 2);
+            } else {
+                copy($tmpFile, $saveFile);
+                Firma::saveIfNotExists($annexe, $fctAl->idSao);
+            }
+            if (file_exists($tmpFile)) {
+                unlink($tmpFile);
+            }
+            $driver->get(self::HTTPS_FOREMP_EDU_GVA_ES_INDEX_PHP_OP_2_SUBOP_0);
+            sleep(1);
+        }
+        return true;
+    }
+    private static function waitForFile($filePath, $timeout)
+    {
+        $startTime = time();
+        while (!file_exists($filePath)) {
+            if (time() - $startTime > $timeout) {
+                throw new \Exception("Timeout waiting for file: $filePath");
+            }
+            sleep(1);
+        }
     }
 }
